malloc: protect stats with lock
[ upstream commit f13d705a619d7a7e04d51b686dd62e890ecf1bd5 ]

When we're gathering statistics, we are traversing the freelist,
which may change under our feet in multithreaded scenario. This
is verified by occasional segfaults when running malloc autotest
on a machine with big amount of cores.

This patch protects malloc heap stats call with a lock. It changes
its definition in the process due to locking invalidating the
const-ness, but this isn't a public API, so that's OK.

Fixes: 2a5c356e177d ("memory: stats for malloc")

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>