net/vdev_netvsc: add check for specifying by 1 way
[ upstream commit 118d2f8096a528b46e1fa5fe2104cf7140278d94 ]

There are now 2 ways to specify a netvsc device by the EAL command
line - either by the interface name or by the MAC address.

The user should not specify a netvsc device using more than 1 way,
Thus, if a device is specified in more than 1 way, the driver stops
to probe it.

Validate it in the driver initialization.

Signed-off-by: Matan Azrad <matan@mellanox.com>