net/vdev_netvsc: prefer netvsc devices in scan
[ upstream commit 4205da3a470974b879d05d337fff2e8bdfec1f70 ]

There is an option to run a non-netvsc device as a netvsc device only
when the "force" parameter is set to 1 in the EAL command line.
Consequently, more than one device may be found to be matching the
"mac" parameter specifying the device.

Prefer netvsc devices to be scanned before any non-netvsc device, even
when the "force" parameter is set.

Signed-off-by: Matan Azrad <matan@mellanox.com>