vhost: check all range is mapped when translating GPAs
There is currently no check done on the length when translating
guest addresses into host virtual addresses. Also, there is no
guanrantee that the guest addresses range is contiguous in
the host virtual address space.

This patch adapts gpa_to_vva()  and its callers to return and
check the mapped size. If the mapped size is smaller than the
requested size, the caller handle it as an error.

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>