net/mlx4: fix inner RSS support for broken kernels
[ upstream commit 828a4ce3c941b2ab3cf6bb65d30339cf8bb2bdf6 ]

Linux 4.15 and 4.16 may report inner RSS as a supported capability of the
device, however it can't be used due to missing code in the kernel.

This triggers an error when creating the default hash QP and prevents this
PMD from starting up without a prior call to rte_flow_isolate().

Fixes: 55e8991e3199 ("net/mlx4: restore inner VXLAN RSS support")

Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>