net/vmxnet3: gather offload data on first and last segment
[ backported from upstream commit 5e5ac26f958f10ef3f1b443ed74c13898cc82e87 ]

Offloads are split between first and last segment of a packet.
Call a single vmxnet3_rx_offload function that will contain all
offload operations. This patch does not introduce any code modification.

Pass a vmxnet3_hw as parameter to the function, it is not presently
used in this patch, but will be later used for TSO offloads.

Signed-off-by: Didier Pallard <didier.pallard@6wind.com>
Acked-by: Yong Wang <yongwang@vmware.com>

Conflicts:
	drivers/net/vmxnet3/vmxnet3_rxtx.c : do not change packet type
	behaviour on stable branch; continue to parse packet
	to return precise RTE_PTYPE_L3_IPV4, rather than returning
	RTE_PTYPE_L3_IPV4_UNKNOWN like in master branch.