app/testpmd: fix forward ports Rx flush
[ upstream commit 47a767b2ee253ccdd084af88d3e390ed9e74f20b ]

A port Rx queue flush is done when the packet forwarding starts in
order to clean the port statistics for a new traffic session.

The flush operation is wrongly called before the update of the new
forward ports, and may fail due to flush operation for an invalid port
configured by the old session.

Move the new forward port setup to be done before the Rx queue flush.

Fixes: 7741e4cf16c0 ("app/testpmd: VMDq and DCB updates")

Signed-off-by: Matan Azrad <matan@mellanox.com>