net/netvsc: resize event buffer as needed
[ upstream commit 1f2766b7ee02c434b554513193c2a2be6664601d ]

The event buffer was changed to be a fixed size value,
had a couple of issues. The big one is that rte_free was still
being called for a pointer that was not setup with rte_malloc().

The event buffer was also too small to handle heavy receive
traffic; and running the event buffer out would crash
the application.

Fix by going back to a dynamically resized event buffer.
And grow it by 25% to avoid lots of realloc's.

Fixes: 530af95a7849 ("bus/vmbus: avoid signalling host on read")

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>