net/mlx5: support missing counter in extended statistics
[ upstream commit 1a611fdaf6ec3d860d7b0f39cdd8b754ef51ebc3 ]

The current code would fail if one of the counters DPDK counters was not
found on the device counters.

As representors and PF port has different counters the both cannot work
together.

Addressing this issue by making the counter init more flexible to
contain all the counter found and skipping the error.

Signed-off-by: Shahaf Shuler <shahafs@mellanox.com>