net/ena: recreate HW IO rings on start and stop
[ upstream commit df238f84c0a21d642bb9517d0c75ba831eeceb46 ]

On the start the driver was refilling all Rx buffs, but the old ones
were not released. That way running start/stop for a few times was
causing device to run out of descriptors.

To fix the issue, IO rings are now being destroyed on stop, and
recreated on start. That way the device is not losing any descriptors.

Furthermore, there was also memory leak for the Rx mbufs, which were
created on start and not destroyed on stop.

Fixes: eb0ef49dd5d5 ("net/ena: add stop and uninit routines")

Signed-off-by: Michal Krawczyk <mk@semihalf.com>