mk: disable gcc AVX512F support
[ upstream commit 8d07c82b239f7782be0d3afbb626390e2cb0c8f5 ]

This is a workaround to prevent a crash, which might be caused by
optimization of newer gcc (7.3.0) on Intel Skylake.

This disables AVX512F support of gcc by adding -mno-avx512f if it is
disabled in DPDK (CONFIG_RTE_ENABLE_AVX512=n).

This does not apply to the meson build as that doesn't have such an option
but always enable AVX512F whenever supported.

Bugzilla ID: 97

Signed-off-by: Yongseok Koh <yskoh@mellanox.com>