examples/l3fwd-power: fix power library fallback
[ upstream commit 6265115f94d8d40fa1e78e8ea01f6de3a369c61f ]

This patch replaces the rte_exit routine with error printing when
init_power_library() fails and by that restores the previous behavior
of the program (which was to issue an error message and continue
working if init_power_library fails). This allows the user to still
experience the Rx interrupts feature of the DPDK demonstrated in
this program.

Fixes: f88e7c175a68 ("examples/l3fwd-power: add high/regular perf cores options")

Signed-off-by: Moti Haimovsky <motih@mellanox.com>
Acked-by: David Hunt <david.hunt@intel.com>