doc: add dependency for PDF in contributing guide
[ upstream commit 63a71926e9f498bc121cec30de1ccae6970da612 ]

There is a missing depenency for building docs with "make doc-guides-pdf".
This causes it to break with "make[3]: latexmk: Command not found". This
was observed and reported in https://bugs.dpdk.org/show_bug.cgi?id=182
This patch fixes this issue by adding the latexmk package dependency in
sub-section 4.3.1 of the contributing guide ("Dependencies").

Bugzilla ID: 182

Signed-off-by: Rami Rosen <ramirose@gmail.com>