event/sw: fix enqueue checks in self-test
[ upstream commit 1b03e2929114d4ebfc7a2159d0d45f0a3e9a7ba6 ]

This patch fixes a number of instances of the same return
value mis-check, where previously we checked for a negative
return value as error, however the API returns an unsigned
integer, so these return value checks are invalid.

The rte_event_enqueue_burst() API returns the number of
events enqueued, so in order to identify the error case,
we must check for != the number of intended enqueues.

Fixes: cd1a9e3eab55 ("test/eventdev: add SW tests for load balancing")

Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>