mk: disable warnings for packed mem config data structure
The rte_mem_config data structure is marked as packed, but we access
members of that structure via pointers in the code. This leads to warnings
with later gcc compilers e.g. gcc9, so disable those warnings to allow
clean builds.

For older GCC versions, which may not have the flag, there is no need to
filter it out as adding unrecognised warning disable flags to GCC is always
safe [Ref: https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html]

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Acked-by: Luca Boccassi <luca.boccassi@microsoft.com>