kni: fix build with kernel 5.3
The include of kni_fifo should be local as kni provides it.
With newer kernels build systems this will run into an issue on
external module compilation.

/var/lib/dkms/dpdk-rte-kni/17.11.6/build/kni_net.c:40:10:
fatal error: kni_fifo.h: No such file or directory
 #include <kni_fifo.h>
          ^~~~~~~~~~~~

Like the includes to kni_dev.h this should be local as well.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Acked-by: Luca Boccassi <bluca@debian.org>