README: indicate default branch is not used
There has been a commit history on the dpdk-stable repo
default branch even though it is not actively used or really
defined what should be there. This is the default branch when
cloning the repo so it can be confusing.

Drop all history on dpdk-stable repo default branch back to the
initial commit and add to the existing README blurb that
stable/LTS branches should be used.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>
Acked-by: Luca Boccassi <bluca@debian.org>
Acked-by: David Marchand <david.marchand@redhat.com>
Acked-by: Thomas Monjalon <thomas@monjalon.net>