doc: remove temporary files and directories
It is possible that someone may have remotes for dpdk
and dpdk-stable in the same directory and build for both.

The temporary files generated in a doc build on dpdk were
interfering with the docs build on dpdk-stable.

In doc-clean remove directories which contain text files
and no .rst files to allow docs to build correctly on
dpdk-stable.

Bugzilla ID: 364

Signed-off-by: Muhammad Bilal <m.bilal@emumba.com>