kni: fix gcc 10 ethtool build error
gcc 10 is smart enough to see that the usec_interval passed into
e1000_phy_has_link_generic() is sometimes a large constant (100000)
and under some conditions it may be passed directly into udelay().

udelay generates an error when it gets these large values by inserting
an undefined function:

if (__builtin_constant_p(n)) {                          \
        if ((n) / 20000 >= 1)                           \
                 __bad_udelay();                        \

And hence we see the error in the form of:
ERROR: "__bad_udelay"
[/root/dpdk/build/kernel/linux/kni/rte_kni.ko] undefined!

Fix this by using similar code to elsewhere in the function whereby
msec_delay_irq() is used when usec value >= 1000.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>
Acked-By: Luca Boccassi <bluca@debian.org>