vhost: remove zero-copy and client mode restriction
Commits [1] were backported as fixes from the latest DPDK release
to the stable branches. They are detecting if vhost-user client and
zero-copy are being used together. As this can be problematic in
some cases, this combination was being rejected by those patches.

It might be ok to deprecate this combination in latest DPDK,
but for stable releases we should not remove the functionality
as it may be in use with earlier releases from the same stable branch.

In fact, we know that this functionality is used at least in OVS in
multiple releases/active branches.

This patch removes the restriction of zero-copy and client mode
and adds a warning for the user.

[1]
commit 81e025d7ed6a
("vhost: prevent zero-copy with incompatible client mode")

commit 0309429b1988
("vhost: fix zero-copy server mode")

Reported-at: https://mails.dpdk.org/archives/stable/2020-June/023562.html
Reported-by: Hariprasad Govindharajan <hariprasad.govindharajan@intel.com>
Signed-off-by: Kevin Traynor <ktraynor@redhat.com>
Acked-by: Luca Boccassi <bluca@debian.org>