net/mlx5: fix PCI address lookup
[ upstream commit 482a1d34b60e7b918a39b2ccd32e06638f8c5aa4 ]

mlx5 PMDs use the mlx5_dev_to_pci_addr() routine to convert
Infiniband device name to the Bus-Device-Function location
on the PCI bus. The routine returned success even in case of
not found identification string. On caller side it likely
caused the wrong match with the BDF of previous device
resulting in wrong representor and master recognitions.

Fixes: 771fa900b73a ("mlx5: introduce new driver for Mellanox ConnectX-4 adapters")

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>