net/mlx5: fix doorbell register mmap offset
The rdma-core might support multiple doorbell registers within
single User Access Region page. The registers beside the first one
are considered as the children ones and it is supposed there is no
extra need to mmap them due the page is already mapped for the first
("parent") register and rdma-core provides uar_map_offset as zero
for the children registers. This caused the DPDK secondary process
crash [1] on mmap() attempt with zero offset parameter.

This patch provides the workaround, to perform correct mapping
in the secondary process we should find the uar_map_offset from
the parent register in the UAR page and use found non-zero one
for mapping.

[1] https://bugs.dpdk.org/show_bug.cgi?id=618

Fixes: 9d2cbd9ea8e7 ("net/mlx5: remove device register remap")

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>