build: fix scheduler macro definition for meson
The make and meson macro formats differ slightly,
the make format is checked when registering the scheduler_autotest
command, and when including dependency on the scheduler pmd for the test,
l2fwd-crypto and test-crypto-perf apps. When building with meson,
these things were not done, and the scheduler autotest was unavailable.

To fix this, the scheduler macro in make format is also defined when
using meson, and the meson build files for the apps use the meson format
when adding the scheduler dependency.

A generalised fix patch [1] for this issue didn't make it into 19.11.6,
but given that does not apply cleanly and introduces other unwanted
changes, this patch is an alternative fix for the scheduler test issue
only.

Fixes: 4dfc61fe367c ("app: fix missing dependencies")
Fixes: dcadbbde8e61 ("crypto/null: build with meson")
Fixes: 52e2991eab28 ("examples/l2fwd-crypto: fix missing dependency")
Cc: bruce.richardson@intel.com
Cc: stable@dpdk.org

Signed-off-by: Ciara Power <ciara.power@intel.com>