doc: fix HiSilicon copyright syntax
[ upstream commit be1650f734b7e478ce485587266e0f858c0cc21f ]

This patch fixes HiSilicon copyright syntax.

According to the suggestion of our legal department,
to standardize the copyright license of our code to
avoid potential copyright risks, we make a unified
modification to the "Hisilicon", which was nonstandard,
in the main modules we maintain.

We change it to "HiSilicon", which is consistent with
the terms used on the following official website:
https://www.hisilicon.com/en/terms-of-use.

Fixes: 565829db8b8f ("net/hns3: add build and doc infrastructure")

Signed-off-by: Min Hu (Connor) <humin29@huawei.com>