net/bnxt: fix auto-negotiation on Whitney+
[ upstream commit 70c7d4184cb9a32ab6c65ce67ce5fbe3a5be5294 ]

Driver should enable autoneg on a port if FW supports it.
Because of a wrong check, driver is not enabling autoneg
on a port after setting forced speed on Whitney+.

There was some dependency on other changes which are not there
in the 19.11 stable. Merged the minimal needed changes for this.

Fixes: 7bc8e9a227cc ("net/bnxt: support async link notification")

Signed-off-by: Kalesh AP <kalesh-anakkur.purayil@broadcom.com>
Reviewed-by: Somnath Kotur <somnath.kotur@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>