test/power: fix CPU frequency when turbo enabled
[ upstream commit 29343b9030e38e8c3519ba01cb66724d45b13dc8 ]

On arm platform, the value in "/sys/.../cpuinfo_cur_freq" may not
be exactly the same as what was set when using CPPC cpufreq driver.
For other cpufreq driver, no need to round it currently, or else
this check will fail with turbo enabled. For example, with acpi_cpufreq,
cpuinfo_cur_freq can be 2401000 which is equal to freqs[0].It should
not be rounded to 2400000.

Backport-Note:
This is a version of the patch for 19.11.9 that fixes this issue
without the dependency of having the CPPC support applied first.

Fixes: 606a234c6d360 ("test/power: round CPU frequency to check")
Cc: stable@dpdk.org

Signed-off-by: David Hunt <david.hunt@intel.com>
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>