eal/freebsd: ignore in-memory option
[ upstream commit 0faa4cfc50275751a86ed17147175c78a36f0da2 ]

The in-memory option is not supported on FreeBSD so print a warning and
ignore the flag when it is specified for BSD apps. The lack of support
is due to the different way in which memory is managed on FreeBSD using
the contigmem driver rather than via a hugetlbfs filesystem.

Fixes: 14de8734c401 ("eal: add --in-memory option")

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>