Revert "net/ena: trigger reset on Tx prepare failure"
This reverts commit 91867d8536914374062693fd5fddc08ae267a41f.
ENA Tx logic in the DPDK v19.11 is a bit different than the one from
which the original commit was backported (DPDK v21.08). As calculations
of the available ring size is not very precise, ena_com_prepare_tx()
function can fail because the ring is full - it's not a fault in the
current logic, so it shouldn't print error logs nor trigger the device
reset.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>