kni: fix build for SLES15-SP3 (Make)
[ upstream commit c28e2165ec360c39ea0e3a0c02b4229966697384 ]

Note: The actual commit c28e2165e is already in 19.11. This is the
backport by Kalesh to also fix make based builds.

As suse version numbering is inconsistent to determine Linux kernel
API to be used. In this patch we check parameter of 'ndo_tx_timeout'
API directly from the kernel source. This is done only for suse build.

Bugzilla ID: 812
Cc: stable@dpdk.org

Tested-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: Kalesh Anakkur Purayil <kalesh-anakkur.purayil@broadcom.com>