igb_uio: fix build for switch fall through
Linux is using '-Wimplicit-fallthrough=5' compiler option, which doesn't
take any fall through comments into account but only uses compiler
'fallthrough' attribute to document fall through action is intended.

"falls through" comment was used in the code which is causing a build
error now, this patch converts comment to the 'fallthrough' macro
defined in the Linux.

To cover the case where Linux version doesn't have the macro, defined it
in the compatibility header too.

Bugzilla-ID: 907

Signed-off-by: Ferruh Yigit <ferruh.yigit@intel.com>
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>