net/ena: fix checksum flag for L4
[ upstream commit b2d2f1cf89a6c3661e68b750364e4a7018f954ba ]

Some HW may invalidly set checksum error bit for the valid L4 checksum.
To avoid drop of the packets in that situation, do not indicate bad
checksum for L4 Rx csum offloads. Instead, set it as unknown, so the
application will re-verify this value.

The statistics counters will still work as previously.

Fixes: 05817057faba ("net/ena: fix indication of bad L4 Rx checksums")

Signed-off-by: Michal Krawczyk <mk@semihalf.com>