Add DPDK PMD to support corundum/mqnic
This initial driver version for the open source NIC or
in-network compute platform, mqnic of the corundum project,
supports a single interface (interface #1 of interfaces #0 - #N)
and a single port per interface.
The driver cannot yet transmit fragmented Frames and also
does not support bursts at this time.

This driver was tested with a 10G version of mqnic build for
a Fidus Sidewinder-100 board. It was attached to a standard intel
NIC (X710 based). Both hosts were running Ubuntu 18.04 with the
intel NIC being operated using the standard Linux driver and the
mqnic being operated by this new DPDK PMD.

Using scapy and tcpdump on the intel NIC side, this driver has
been validated, while the mqnic side was running either the
"dpdk-testpmd" oder "l2fwd" example applications:

  intel side:
    sendp(Ether(src='01:00:0c:cc:cc:cc', dst='00:11:22:33:44:55')
          /IP(src="1.2.3.4", dst="3.4.5.6")/UDP(dport=9)
          /b"hello, world!", iface="ens4f0")

  mqnic side:
    dpdk-l2fwd -l 0 -- -p 0x1

This driver is built against
	- DPDK-Stable v20.11.3 [1] as a base for the DPDK driver
        - a specific, partially outdated Corundum/mqnic version [2]
          as the base for the mqnic adapted from the kernel driver
          and for the hardware design

[1] http://git.dpdk.org/dpdk-stable/commit/?h=v20.11.3&id=cfcca6990363250939dd92962b3bd3c31b743c75
[2] https://github.com/corundum/corundum/commit/0965c77b8e787f56bed84d539821fec448cd24a3

Signed-off-by: Bjoern Petersen <bjoern.petersen@missinglinkelectronics.com>
Signed-off-by: Ulrich Langenbach <ulrich.langenbach@missinglinkelectronics.com>