net/mlx5: fix netlink header path
Use netlink.h provided by the kernel headers instead
of adding a dependency on the headers provided by the libnl package,
which weren't being found in some cases, even with libnl-3 and
libnl-route-3 libs installed via the package manager.
Also see usage in netlink(3).

Resolves the build error:
  drivers/net/mlx5/mlx5.h:15:10:
    fatal error: netlink/netlink.h: No such file or directory
The error reproduces on Ubuntu 16.04/18.04/20.04, openSUSE Leap 15.3,
and CentOS 7/8. It reproduces with Make and Meson.

Fixes: 7bd18f45ed47 ("common/mlx5: add Netlink event helpers")
Cc: dkozlyuk@nvidia.com

Signed-off-by: Ali Alnubani <alialnu@nvidia.com>