net/mlx5: cleanup Netlink interrupt handler
[ upstream commit 17f95513adca1f4ee5bba305b154ac984ee50cee ]

When the upstream commit was backported by the offending commit,
the cleanup part of the fix was missed. The interrupt handler
was not unregistered and the Netlink socket was not closed
with the device closing. The socket was leaking and the handler
could continue working with random content in the freed memory:

    net_mlx5: Failed to process Netlink events: \
        Socket operation on non-socket

Fixes: edea6144b47f ("net/mlx5: fix link status change detection")

Signed-off-by: Dmitry Kozlyuk <dkozlyuk@nvidia.com>
Acked-by: Suanming Mou <suanmingm@nvidia.com>