drivers/net/mqnic: Add support for corundum project mqnic
This driver was tested with a 10G version of mqnic build for
an AMD ZCU106 board. It was attached to a standard intel
NIC (X710 based). Both hosts were running Ubuntu 18.04 with the
intel NIC being operated using the standard Linux driver and the
mqnic being operated by this new DPDK PMD.

Using a packet generator like pktgen and tcpdump on the intel NIC side, this driver has
been validated, while the mqnic side was running either the
"dpdk-testpmd" or "l2fwd" example applications with no-mac-updating option enabled:

This driver is built against
        - DPDK-Stable v20.11.3 [1] as a base for the DPDK driver
        - a specific, partially outdated Corundum/mqnic version [2]
          as the base for the mqnic adapted from the kernel driver
          and for the hardware design

[1] http://git.dpdk.org/dpdk-stable/commit/?h=v20.11.3&id=5b8fb20bada486ce43603530659533fab7130a55
[2] https://github.com/corundum/corundum/commit/0965c77b8e787f56bed84d539821fec448cd24a3

Signed-off-by: Björn Petersen <bjoern.petersen@missinglinkelectronics.com>
Signed-off-by: Ulrich Langenbach <ulrich.langenbach@missinglinkelectronics.com>